User Id now accessible by token

+++ backend/springCrudAPI/src/main/java/com/example/springBoot/controller/UserController.java
@@ -23,57 +23,85 @@ import java.util.List;
 public class UserController {
 	
 	@Autowired //Injected user repository annotation
-	private UserRepository userRepository; //using all the CRUD database operations from userRepository interface
-	
-	// The followings are the rest APIs that we will be creating next
+	private UserRepository userRepository;
 	
 	//Get all users
 	@GetMapping
-	public List<User> getAllUsers(){ //This API returns a list of all users
+	public List<User> getAllUsers(){
 		return this.userRepository.findAll();
 	}
-	//Get all users by ID
+	
+	//Get the user by ID
 	@GetMapping("/{id}")
 	public User getUserByID(@PathVariable(value = "id") long userId) { 
-//The path variable gets it from database
-		System.out.println("From get method" + userId);
 		
-		//User user = new User();
-		
-		//return  user;
-	
-		return this.userRepository.findById(userId).
+		if(this.userRepository.existsById(userId)) {
+			//The path variable gets it from database
+			return this.userRepository.findById(userId).
 				orElseThrow(() -> new ResourceNotFoundException("User not found with the ID: " + userId));
-		
+		}
+		else {
+			throw new ResourceNotFoundException("User not found with the ID: " + userId);
+		}
 	}
+	
 	//Create user
 	@PostMapping
 	public User createUser(@RequestBody User user) {
+		String tk = user.generateToken(user.getID());
+		user.setToken(tk);
 		return this.userRepository.save(user);
 		
 	}
+	
 	//Update user
 	@PutMapping("/{id}")
 	public User updateUser(@RequestBody User user, @PathVariable("id") long userId) {
-		User existingUserObj =  this.userRepository.findById(userId).
-				orElseThrow(() -> new ResourceNotFoundException("User not found with the ID: " + userId));
 		
-		existingUserObj.setFirstName(user.getFirstName());
-		existingUserObj.setLastName(user.getLastName());
-		existingUserObj.setEmail(user.getEmail());
-		existingUserObj.setToken(user.getToken());
-		existingUserObj.setPwd(user.getPwd());
-		existingUserObj.setRole(user.getRole());
-		
-		return this.userRepository.save(existingUserObj);	
-		
+		if(this.userRepository.existsById(userId)) {
+			User existingUserObj =  this.userRepository.findById(userId).
+					orElseThrow(() -> new ResourceNotFoundException("User not found with the ID: " + userId));
+			
+			existingUserObj.setFirstName(user.getFirstName());
+			existingUserObj.setLastName(user.getLastName());
+			existingUserObj.setEmail(user.getEmail());
+			existingUserObj.setPwd(user.getPwd());
+			
+			return this.userRepository.save(existingUserObj);	
+			
+		}
+		else {
+			throw new ResourceNotFoundException("User not found with the ID: " + userId);
+		}	
 	}
+	
 	//Delete user by ID 
 	@DeleteMapping("/{id}/")
 	public ResponseEntity<User> deleteUser(@PathVariable("id") long userId){
-		User existingUserObj =  this.userRepository.findById(userId).
-				orElseThrow(() -> new ResourceNotFoundException("User not found with the ID: " + userId));
-		this.userRepository.delete(existingUserObj);
-		return ResponseEntity.ok().build();
+		
+		if(this.userRepository.existsById(userId)) {
+			User existingUserObj = this.userRepository.findById(userId).
+					orElseThrow(() -> new ResourceNotFoundException("User not found with the ID: " + userId));
+			this.userRepository.delete(existingUserObj);
+			return ResponseEntity.ok().build();
+		}
+		else {
+			throw new ResourceNotFoundException("User not found with the ID: " + userId);
+		}
+	}
+	
+	//Get userId by Token
+	@GetMapping("/{Token}")
+	public String getIdByToken(@RequestBody User user, @PathVariable("Token") String Token){
+		String userId = "";
+		try {
+			User user2 = userRepository.searchByToken(Token);
+			userId = String.valueOf(user2.getID());
+			
+		}
+		catch(Exception e) {
+			return "User does not exist";
+		}
+		return "User ID: " + userId;
 	}
 }
